feat(language/compiler/check): equate VariableRole::Parameter slots in constrain_structural + reject SignatureMatch residual failures

+++ language/compiler/src/check/tests/expressions/pattern/variant.rs
@@ -589,7 +589,7 @@ function describe(mode: Mode): int32 {
     match (mode) {
         Mode.Read => 10
         Mode.Write => 20
-    } as int32
+    }
 }
 
 === checked ===
